[SPARK-42815][SQL] Subexpression elimination support shortcut expression
### What changes were proposed in this pull request?

Add a new config to shortcut subexpression elimination for expression `and`, `or`.

The subexpression may not need to eval even if it appears more than once.
e.g., `if(or(a, and(b, b)))`, the expression `b` would be skipped if `a` is true.

### Why are the changes needed?

avoid eval unnecessary expression.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

add test

Closes #40446 from ulysses-you/shortcut.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>