[SPARK-42832][SQL] Remove repartition if it is the child of LocalLimit
### What changes were proposed in this pull request?

This PR enhances `CollapseRepartition` to remove repartition if it is the child of `LocalLimit`. Because its output is determined by the number of partitions and the expressions of the Repartition. Therefore, it is feasible to remove Repartition except for repartition by nondeterministic expressions, because users may expect to randomly take data.
For example:
```sql
SELECT /*+ REBALANCE */ * FROM t WHERE id > 1 LIMIT 5;
```

Before this PR:
```
== Optimized Logical Plan ==
GlobalLimit 5
+- LocalLimit 5
   +- RebalancePartitions
      +- Filter (isnotnull(id#0L) AND (id#0L > 1))
         +- Relation spark_catalog.default.t[id#0L] parquet
```

After this PR:
```
== Optimized Logical Plan ==
GlobalLimit 5
+- LocalLimit 5
   +- Filter (isnotnull(id#0L) AND (id#0L > 1))
      +- Relation spark_catalog.default.t[id#0L] parquet
```

Note that we don't remove repartition if it looks like the user might want to take data randomly. For example:
```sql
SELECT /*+ REPARTITION(3) */ * FROM t WHERE id > 1 LIMIT 5;
SELECT * FROM t WHERE id > 1 DISTRIBUTE BY random() LIMIT 5;
```

### Why are the changes needed?

Reduce shuffle to improve query performance. The use case is that we add a repartition to improve the parallelism on a JDBC table:
<img src="https://user-images.githubusercontent.com/5399861/225855582-c3c81c7d-4617-4104-b669-76749a7468a0.png" width="400" height="700">

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Unit test.

Closes #40462 from wangyum/SPARK-42832.

Authored-by: Yuming Wang <yumwang@ebay.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>