[SPARK-42969][CONNECT][TESTS] Fix the comparison the result with Arrow optimization enabled/disabled
### What changes were proposed in this pull request?

Fixes the comparison the result with Arrow optimization enabled/disabled.

### Why are the changes needed?

in `test_arrow`, there are a bunch of comparison between DataFrames with Arrow optimization enabled/disabled.

These should be fixed to compare with the expected values so that it can be reusable for Spark Connect parity tests.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Updated the tests.

Closes #40612 from ueshin/issues/SPARK-42969/test_arrow.

Authored-by: Takuya UESHIN <ueshin@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>