[SPARK-43019][SQL] Move Ordering to PhysicalDataType
### What changes were proposed in this pull request?

This PR proposes that we start to move ordering to PhysicalDataType. This is to simplify the DataType class to make it become a simple interface without coupling too many internal representations.

### Why are the changes needed?

To make DataType become a simpler interface, non-public code can be moved outside of the DataType class.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

UT

Closes #40651 from amaliujia/catalyst_datatype_refactor.

Authored-by: Rui Wang <rui.wang@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>