[SPARK-41537][INFRA][CONNECT][FOLLOW-UP] Removes breaking changes within master branch
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/39294 that proposes to remove the breaking change detection within `master` branch.

### Why are the changes needed?

Apache Spark does not guarantee the compatibility between individual commits but between releases according to semantic versioning.

It would have been great if we can use this job as a sort of warning but this feature is missing in GitHub Actions. See also:
- https://github.com/orgs/community/discussions/15452
- https://github.com/actions/runner/issues/2347

Therefore, we cannot use this as a warning.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Will be tested within my fork first.

Closes #40673 from HyukjinKwon/SPARK-41537-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>