[MINOR][SQL][TESTS] Tests in `SubquerySuite` should not drop view created in `beforeAll`
### What changes were proposed in this pull request?

Change tests to avoid replacing and dropping a temporary view that is created in `SubquerySuite#beforeAll`.

### Why are the changes needed?

When I added a test for SPARK-42937, it tried to use the view `t`, which is created in `beforeAll`. But because other tests would replace and drop this view, the new test would fail. As a result, that new test had to re-create `t` from scratch.

This change will allow `t` to be used by new tests.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests.

Closes #40717 from bersprockets/bad_drop_view.

Authored-by: Bruce Robbins <bersprockets@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>