[SPARK-43124][SQL] Dataset.show projects CommandResults locally
### What changes were proposed in this pull request?
`DataSet.show()` currently triggers a job for a simple `show tables` command. This is because the command output contains an `isTemporary` boolean column that needs to be casted to string when we use `show()` on the dataset.
This PR converts `CommandResult` to `LocalRelation` and let `ConvertToLocalRelation` to do the casting locally to avoid triggering job execution.

### Why are the changes needed?
A simple `show tables` shouldn not require an executor.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Added new UT.

Closes #40779 from peter-toth/SPARK-43124-dataset-show-projects-commandresults-locally.

Authored-by: Peter Toth <peter.toth@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>