[SPARK-43204][SQL] Align MERGE assignments with table attributes
### What changes were proposed in this pull request?

This PR extends `ResolveRowLevelCommandAssignments` to also cover MERGE assignments.

### Why are the changes needed?

Similar to SPARK-42151, these changes are needed so that we can rewrite MERGE statements into executable plans for tables that support row-level operations. In particular, our row-level mutation framework assumes Spark is responsible for building an updated version of each affected row and that row is passed back to the data source.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR comes with tests.

Closes #40919 from aokolnychyi/spark-43204.

Authored-by: aokolnychyi <aokolnychyi@apple.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>