[SPARK-42419][FOLLOWUP][CONNECT][PYTHON] Remove unused exception
### What changes were proposed in this pull request?

This is follow-up for https://github.com/apache/spark/pull/39991 to remove unused exception.

### Why are the changes needed?

`PySparkTypeError` never raises because we checked the type already in the codes above `if type(startPos) != type(length):` and raise `PySparkTypeError` for unsupported type for `length`.

### Does this PR introduce _any_ user-facing change?

No. it's minor code cleanup

### How was this patch tested?

The existing CI should pass.

Closes #40927 from itholic/connect_col_followup.

Authored-by: itholic <haejoon.lee@databricks.com>
Signed-off-by: Ruifeng Zheng <ruifengz@apache.org>