[SPARK-43136][CONNECT] Adding groupByKey + mapGroup + coGroup functions
### What changes were proposed in this pull request?
This PR adds the mapGroup and coGroup support for Scala client.
The client exposes the same SQL API to users, but uses an internal impl of the `KeyValueGroupedDatasetImpl` to keep track of initial types of the grouping keys and values. This is needed to support `keyAs` and `mapValues` type modifications to the dataset.

It adds one the sorting expression into the proto so that the sorting expressions can be passed to the server.

### Why are the changes needed?
API completion.

### Does this PR introduce _any_ user-facing change?
Add new Scala Client `KeyValueGroupedDataset` APIs.

### How was this patch tested?
Client E2E tests that only work for SBT due to UDFUtil not synced to the server with maven build.
The maven tests will be address in a follow up [PR](https://github.com/apache/spark/pull/40762)

Closes #40729 from zhenlineo/key-value-grouped.

Authored-by: Zhen Li <zhenlineo@users.noreply.github.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>