[SPARK-43293][SQL] `__qualified_access_only` should be ignored in normal columns
### What changes were proposed in this pull request?

This is a followup of https://github.com/apache/spark/pull/39596 to fix more corner cases. It ignores the special column flag that requires qualified access for normal output attributes, as the flag should be effective only to metadata columns.

### Why are the changes needed?

It's very hard to make sure that we don't leak the special column flag. Since the bug has been in the Spark release for a while, there may be tables created with CTAS and the table schema contains the special flag.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

new analysis test

Closes #40961 from cloud-fan/col.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>