[SPARK-16484][SQL] Use 8-bit registers for representing DataSketches
### What changes were proposed in this pull request?

This is a follow-up PR to #40615.

In this PR we make sure that all sketches produces by the new expressions `hll_sketch_agg`, `hll_union_agg`, and `hll_union` are actually HLL DataSketches represented with 8-bit registers.

Using 8-bit registers improves the performance of the expressions at the expense of using more memory (the current implementation uses 4-bit registers by default, which is the default choice in the DataSketches library).

Note that the register size of the DataSketches HLL sketch does not have any impact on the accuracy of the sketch.

### Why are the changes needed?

To improve the performance of the new HLL-related expressions.

### Does this PR introduce _any_ user-facing change?

No in terms of SQL surface. However, the results of the `hll_sketch_agg`, `hll_union_agg`, and `hll_union` expressions can be different since we are changing the sketch representation. However, since these expressions have not been released yet we are not facing any breaking changes.

### How was this patch tested?

Existing tests suffice.

Closes #41021 from mkaravel/datasketches-change-register-size.

Authored-by: Menelaos Karavelas <menelaos.karavelas@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>