[SPARK-43331][CONNECT] Add Spark Connect SparkSession.interruptAll
### What changes were proposed in this pull request?

Currently, queries that are ran using Spark Connect cannot be interrupted. Even when the RPC connection got broken, the Spark jobs on the server continue running.

This PR proposes a
```
rpc Interrupt(InterruptRequest) returns (InterruptResponse) {}
```
server RPC API, that can be called from the client as `SparkSession.interruptAll()` to interrupt all actively running Spark Jobs from ExecutePlan executions. In most user scenarios, SparkSession is not used for multiple executions concurrently, but is used sequentially, so `interruptAll()` should serve a big chunk of user needs. It can also be used to clean up.

To keep track of executions, we introduce `ExecutionHolder` to hold the execution state, and make `SessionHolder` keep track of the executions currently running in the session. In this first PR, the interrupt only interrupts running Spark Jobs. As such, it is to a degree best effort, because it will not interrupt commands that don't run Spark Jobs, or it will not interrupt anything if a Spark Job is not running when it the interrupt is received by the server, and  the command will continue running and may continue launching more Spark jobs later in that case.

Future work I plan to design and work on will involve:
* Interrupting any execution. This will involve moving execution from the GRPC handler thread handling ExecutePlan to launching it in a separate thread that can be interrupted. `ExecutionHolder`
* Interrupting executions selectively. This will involve exposing the operationId to the user.
* (Refactor) some cleanup needed around using SessionHolder. Currently, sessionId, userId, various session data is passed around separately. SessionHolder may be passed instead, and then also be extended with more useful APIs for management of the session.

### Why are the changes needed?

Need to have APIs to be able to interrupt running queries in Spark Connect.

### Does this PR introduce _any_ user-facing change?

Yes.
Users of Spark Connect can now call `interruptAll()` on client `SparkSession` object, to send an interrupt RPC to the server, which will interrupt the running queries.

In followup PRs, this will be extended to Python client, and to work not only for interrupting Spark Jobs.

### How was this patch tested?

Added E2E tests to ClientE2ETestSuite for scala connect client.
Added unit tests to test_client for python connect client.

Closes #41005 from juliuszsompolski/connect-cancelall.

Authored-by: Juliusz Sompolski <julek@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>