[SPARK-43250][SQL] Replace the error class `_LEGACY_ERROR_TEMP_2014` with an internal error
### What changes were proposed in this pull request?
In this PR I propose to replace the legacy error class `_LEGACY_ERROR_TEMP_2014`, added as an `IllegalArgumentException` to avoid non-exhaustive case-matching in #22014, with an internal error as it is not triggered by the user space.

### Why are the changes needed?
As the error is not triggered by the user space, the legacy error class can be replaced by an internal error.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
No new tests were added and no tests needed changing because of the nature of the updated error class.

Closes #40951 from amousavigourabi/add-new-function-internal-error.

Lead-authored-by: amousavigourabi <28668597+amousavigourabi@users.noreply.github.com>
Co-authored-by: Atour <28668597+amousavigourabi@users.noreply.github.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>