[SPARK-40912][CORE] Overhead of Exceptions in KryoDeserializationStream
### What changes were proposed in this pull request?
This PR avoid exceptions in the implementation of KryoDeserializationStream.

### Why are the changes needed?
Using an exceptions for end of stream is slow, especially for small streams. It also problematic as it the exception caught in the KryoDeserializationStream could also be caused by corrupt data which would just be ignored in the current implementation.

### Does this PR introduce _any_ user-facing change?
Yes, it changes so some method on KryoDeserializationStream no longer raises EOFException.

### How was this patch tested?
Existing tests.

This PR only changes KryoDeserializationStream as a proof of concept. If this is the direction we want to go we should probably change DerserializationStream isntead so that the interface is consistent.

Closes #38428 from eejbyfeldt/SPARK-40912.

Authored-by: Emil Ejbyfeldt <eejbyfeldt@liveintent.com>
Signed-off-by: Sean Owen <srowen@gmail.com>