[SPARK-43412][PYTHON][CONNECT] Introduce `SQL_ARROW_BATCHED_UDF` EvalType for Arrow-optimized Python UDFs
### What changes were proposed in this pull request?
Introduce `SQL_ARROW_BATCHED_UDF` EvalType for Arrow-optimized Python UDFs.

An EvalType is used to uniquely identify a UDF type in PySpark.

### Why are the changes needed?
We are about to improve nested non-atomic input/output support of an Arrow-optimized Python UDF.

However, currently, it shares the same EvalType with a pickled Python UDF, but the same implementation with a Pandas UDF.

Introducing an EvalType enables isolating the changes to Arrow-optimized Python UDFs.

The PR is also a pre-requisite for registering an Arrow-optimized Python UDF.

### Does this PR introduce _any_ user-facing change?
No user-facing behavior/result changes for Arrow-optimized Python UDFs.

An `evalType`, as an attribute mainly designed for internal use, is changed as shown below:

```py
>>> udf(lambda x: str(x), useArrow=True).evalType == PythonEvalType.SQL_ARROW_BATCHED_UDF
True

# whereas

>>> udf(lambda x: str(x), useArrow=False).evalType == PythonEvalType.SQL_BATCHED_UDF
True
```

### How was this patch tested?
A new unit test `test_eval_type` and existing tests.

Closes #41053 from xinrong-meng/evalTypeArrowPyUDF.

Authored-by: Xinrong Meng <xinrong@apache.org>
Signed-off-by: Xinrong Meng <xinrong@apache.org>