[SPARK-38467][CORE] Use error classes in org.apache.spark.memory
### What changes were proposed in this pull request?
This PR aims to change exceptions created in package org.apache.spark.memory to use error class. This also adds an error class INTERNAL_ERROR_MEMORY and uses that for the internal errors in the package.

### Why are the changes needed?
This is to move exceptions created in package org.apache.spark.memory to error class.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing unit tests.

Closes #41121 from bozhang2820/spark-38467.

Authored-by: Bo Zhang <bo.zhang@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>