[SPARK-43542][SS] Define a new error class and apply for the case where streaming query fails due to concurrent run of streaming query with same checkpoint
### What changes were proposed in this pull request?

We are migrating to a new error framework in order to surface errors in a friendlier way to customers. This PR defines a new error class specifically for when there are concurrent updates to the log for the same batch ID

### Why are the changes needed?

This gives more information to customers, and allows them to filter in a better way

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

There is an existing test to check the error message upon this condition. Because we are only changing the error type, and not the error message, this test is sufficient.

Closes #41205 from ericm-db/SC-130782.

Authored-by: Eric Marnadi <eric.marnadi@databricks.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>