[SPARK-43775][SQL] DataSource V2: Allow representing updates as deletes and inserts
### What changes were proposed in this pull request?

This PR adds a way for data sources to request Spark to represent updates as deletes and inserts.

### Why are the changes needed?

It may be beneficial for data sources to represent updates as deletes and inserts for delta-based implementations. Specifically, it may help to properly distribute and order records before writing.

Delete records set only row ID and metadata attributes. Update records set data, row ID, metadata attributes. Insert records set only data attributes.

For instance, a data source may rely on a metadata column `_row_id` (synthetic internally generated) to identify the row and may be partitioned by `bucket(product_id)`. Splitting updates into inserts and deletes would allow data sources to cluster all update and insert records in MERGE for the same partition into a single task. Otherwise, the clustering key for updates and inserts will be different (inserts will always have `_row_id` as null as it is a metadata column). The new functionality is critical to reduce the number of generated files. It also makes sense in UPDATE operations if the original and new partition of a record do not match.

### Does this PR introduce _any_ user-facing change?

This PR adds a new method to `SupportsDelta` but the change is backward compatible.

### How was this patch tested?

This PR comes with tests.

Closes #41300 from aokolnychyi/spark-43775.

Authored-by: aokolnychyi <aokolnychyi@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>